Working change of chatrooms
Input stream on client now is read into two separate streams: message
that represent message to be shown on chat field and checked constantly;
and management that represent data which is required once.
Refactored and enhanced server ClientWorker.
Merged GUI and ChatClient class into one with better readability.
Overall changed visibility of functions and variables to as strict as
required.